Add Sega Masters System version
I've added the SMS version. Both the Europe and Europe beta. 
I did not include the Europe beta in the html as I didn't think it was as important. But did want to include it in case someone did need it.

SMS doesn't have any line endings so I had to exclude those when comparing to the SMS version.